Add Handler interface
